pyproject.toml
[tool.poetry]
name = "youtube-analytics-project"
version = "0.1.0"
description = ""
authors = ["Evgeny Safonov <jeka-luda@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
isodate = "^0.6.1"


[tool.poetry.group.dev.dependencies]
google-api-python-client = "^2.108.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
